+++ tang-tool/tang-generator/src/main/java/com/tang/generator/utils/VelocityUtils.java
@@ -20,21 +20,26 @@ import com.tang.generator.enumeration.LanguageType;
 import com.tang.generator.enumeration.OrmType;
 
 import static com.tang.commons.utils.StringUtils.format;
-import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_CONTROLLER;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_ENTITY;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_MAPPER;
-import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_CONTROLLER;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_SERVICE;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_SERVICE_IMPL;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_CONTROLLER;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATISFLEX_ENTITY;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATISFLEX_MAPPER;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATISFLEX_SERVICE;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATISFLEX_SERVICE_IMPL;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATISFLEX_CONTROLLER;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_ENTITY;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_MAPPER;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_SERVICE;
 import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_SERVICE_IMPL;
-import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_SERVICE;
-import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_SERVICE_IMPL;
-import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_CONTROLLER;
+import static com.tang.generator.constants.VmTemplate.VM_JAVA_MYBATIS_PLUS_CONTROLLER;
 import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_ENTITY;
 import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_MAPPER;
 import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_SERVICE;
 import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_SERVICE_IMPL;
+import static com.tang.generator.constants.VmTemplate.VM_KOTLIN_MYBATIS_CONTROLLER;
 import static com.tang.generator.constants.VmTemplate.VM_SQL_MENU;
 import static com.tang.generator.constants.VmTemplate.VM_VUE_INDEX;
 import static com.tang.generator.constants.VmTemplate.VM_VUE_INDEX_TS;
@@ -230,6 +235,7 @@ public class VelocityUtils {
         return switch (languageType) {
             case JAVA -> switch (ormType) {
                 case MYBATIS -> List.of(VM_JAVA_MYBATIS_ENTITY, VM_JAVA_MYBATIS_MAPPER, VM_JAVA_MYBATIS_SERVICE, VM_JAVA_MYBATIS_SERVICE_IMPL, VM_JAVA_MYBATIS_CONTROLLER, VM_XML_MYBATIS_MAPPER);
+                case MYBATIS_FLEX -> List.of(VM_JAVA_MYBATISFLEX_ENTITY, VM_JAVA_MYBATISFLEX_MAPPER, VM_JAVA_MYBATISFLEX_SERVICE, VM_JAVA_MYBATISFLEX_SERVICE_IMPL, VM_JAVA_MYBATISFLEX_CONTROLLER);
                 case MYBATIS_PLUS -> List.of(VM_JAVA_MYBATIS_PLUS_ENTITY, VM_JAVA_MYBATIS_PLUS_MAPPER, VM_JAVA_MYBATIS_PLUS_SERVICE, VM_JAVA_MYBATIS_PLUS_SERVICE_IMPL, VM_JAVA_MYBATIS_PLUS_CONTROLLER, VM_XML_MYBATIS_PLUS_MAPPER);
                 default -> Collections.emptyList();
             };
@@ -268,11 +274,11 @@ public class VelocityUtils {
         var apiPath = format("{}/api/{}", VUE_PATH, moduleName);
 
         fileName = switch (template) {
-            case VM_JAVA_MYBATIS_ENTITY, VM_JAVA_MYBATIS_PLUS_ENTITY -> format("{}/entity/{}.java", javaPath, className);
-            case VM_JAVA_MYBATIS_MAPPER, VM_JAVA_MYBATIS_PLUS_MAPPER -> format("{}/mapper/{}Mapper.java", javaPath, className);
-            case VM_JAVA_MYBATIS_SERVICE, VM_JAVA_MYBATIS_PLUS_SERVICE -> format("{}/service/{}Service.java", javaPath, className);
-            case VM_JAVA_MYBATIS_SERVICE_IMPL, VM_JAVA_MYBATIS_PLUS_SERVICE_IMPL -> format("{}/service/impl/{}ServiceImpl.java", javaPath, className);
-            case VM_JAVA_MYBATIS_CONTROLLER, VM_JAVA_MYBATIS_PLUS_CONTROLLER -> format("{}/controller/{}Controller.java", javaPath, className);
+            case VM_JAVA_MYBATIS_ENTITY, VM_JAVA_MYBATISFLEX_ENTITY, VM_JAVA_MYBATIS_PLUS_ENTITY -> format("{}/entity/{}.java", javaPath, className);
+            case VM_JAVA_MYBATIS_MAPPER, VM_JAVA_MYBATISFLEX_MAPPER, VM_JAVA_MYBATIS_PLUS_MAPPER -> format("{}/mapper/{}Mapper.java", javaPath, className);
+            case VM_JAVA_MYBATIS_SERVICE, VM_JAVA_MYBATISFLEX_SERVICE, VM_JAVA_MYBATIS_PLUS_SERVICE -> format("{}/service/{}Service.java", javaPath, className);
+            case VM_JAVA_MYBATIS_SERVICE_IMPL, VM_JAVA_MYBATISFLEX_SERVICE_IMPL, VM_JAVA_MYBATIS_PLUS_SERVICE_IMPL -> format("{}/service/impl/{}ServiceImpl.java", javaPath, className);
+            case VM_JAVA_MYBATIS_CONTROLLER, VM_JAVA_MYBATISFLEX_CONTROLLER, VM_JAVA_MYBATIS_PLUS_CONTROLLER -> format("{}/controller/{}Controller.java", javaPath, className);
             case VM_KOTLIN_MYBATIS_ENTITY -> format("{}/entity/{}.kt", kotlinPath, className);
             case VM_KOTLIN_MYBATIS_MAPPER -> format("{}/mapper/{}Mapper.kt", kotlinPath, className);
             case VM_KOTLIN_MYBATIS_SERVICE -> format("{}/service/{}Service.kt", kotlinPath, className);