+++ tang-extension/tang-mybatis-flex/src/main/kotlin/com/tang/extension/mybatisflex/utils/If.kt
@@ -1,0 +1,35 @@
+package com.tang.extension.mybatisflex.utils
+
+import java.util.Objects
+
+/**
+ * @author Tang
+ */
+object If {
+
+    /**
+     * 判断是否为空
+     */
+    @JvmStatic
+    fun isEmpty(any: Any?): Boolean {
+        if (Objects.isNull(any)) {
+            return true
+        }
+        return when (any) {
+            is String -> any.isEmpty()
+            is Array<*> -> any.isEmpty()
+            is Map<*, *> -> any.isEmpty()
+            is Collection<*> -> any.isEmpty()
+            else -> false
+        }
+    }
+
+    /**
+     * 判断是否非空
+     */
+    @JvmStatic
+    fun nonEmpty(any: Any?): Boolean {
+        return !isEmpty(any)
+    }
+
+}