+++ CHANGELOG.md
@@ -29,3 +29,4 @@
  * Support java 8 time type
  * Support mybatis flex
  * Support mybatis flex order query
+ * Support mybatis flex generate query condition