+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### Dependencies
 
- * Bump spring-boot-dependencies from 3.2.3 to 3.2.4
+ * Bump spring-boot-dependencies from 3.2.3 to 3.2.5
  * Bump spring-boot-maven-plugin from 3.2.3 to 3.2.4
  * Bump jackson from 2.16.1 to 2.17.0
  * Bump redis from 7.2.3 to 7.2.4