+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 ### Dependencies
 
  * Bump spring-boot-dependencies from 3.2.3 to 3.2.5
- * Bump spring-boot-maven-plugin from 3.2.3 to 3.2.4
+ * Bump spring-boot-maven-plugin from 3.2.3 to 3.2.5
  * Bump jackson from 2.16.1 to 2.17.0
  * Bump redis from 7.2.3 to 7.2.4
  * Add mybatis 3.5.16